fixed a warning on crayc++

+++ ugbase/common/util/metaprogramming_util.h
@@ -126,7 +126,7 @@ struct Factorial<1>
  * ---------
  * k! (n-k)!
  */
-template <size_t n, size_t k>
+template <size_t n, int k>
 struct BinomialCoefficient
 {
     enum { value = 	Factorial<n>::value/
